link view history buttons to view history codes
the view history action in the patient table will lead to a working "log," which has the desired patient's information that is displayed on the table.

however, we want this feature to work on the view patient history button on the video upload page. right now, it leads to the history page, but no patient information is being displayed.

i suspect that we have to bring over the entire patient table page into the view patient history button (black)